Predict Sales of Product
Using the prices of the other brands of a certain food item, I created several algorithms to predict the quantity sold of the various brands. This was done using regression. The predicted weights were then input into an excel workbook, and one can test setting the prices of the various brands at different prices to see what will maximize profit. Due to a confidentiality agreement, the food item is not revealed, and the various brands are denoted by names of Pokemon.