Excel Workbook Deliverable
Enter the price that the product was bought at from the manufacturer and the price that the retailer intends to sell it for for the top five brands of the product. Then just hit the VBA buttons to see how much of each brand you would sell to maximize profit.