Drake Relays Alcohol A/B Testing
This research was conducted in my marketing analytics course and shared with Drake Athletics. This document includes a survey that we distributed to Drake students over the age of 21. The purpose of this study was to gauge the extent to which serving alcohol at the Drake Relays would increase attendance (if at all), and thereby increase profits. This write-up includes the thought process behind the survey design, the SAS code which ultimately determined that alcohol was a significant predictor of attendance, and final conclusions.