Car Crashes
In this project, it was our goal to determine what road factors are most likely to cause sever car crashes in the Des Moines Metropolitan Area. This folder includes 1. a link to Tableau visualizations which help identify the most dangerous intersections 2. Significance Testing which determined the most dangerous road conditions and 3. A result of our findings in powerpoint format.