Description
Using only three columns as predictors, I created a decision tree to determine whether someone was likely to pay back their loans / have good credit. I binned several variables to see if that enhanced my accuracy and computed entropy so that the attribute with the highest information gain was selected as the root of the tree.