Description
In this project which was presented to Drake Athletics staff, I sought to find whether selling alcohol at the Drake Relays would increase profit for the event. This required constructing two surveys and distributing it to Drake students over the age of 21, and determining whether alcohol was a significant predictor of attendance using a regression model in SAS. More details can be found in the document.