Description
In this project for marketing analytics, a bank was making phone calls to prospective clients and losing money, as each call cost $5 to make. In this project, using regression techniques, I identify variables that are important to whether a prospective client will sign up with a bank (and therefore make money), and determine which clients should be called.